requirements.txt
fastapi
python-telegram-bot
instaloader
uvicorn
